sample-code/image-processing/thumbnail-generator/requirements.txt
fdk
oci
requests
Pillow
python-magic
